requirements.txt
pandas>=2.0
openpyxl>=3.1
groq>=1.0
streamlit>=1.30
numpy>=1.26
python-dotenv>=1.0
